refactor: move cache safety check into django_cache.py and apply review feedback (#978)
