+++ dmr/throttling/backends/django_cache.py
@@ -1,7 +1,9 @@
 import dataclasses
-from typing import TYPE_CHECKING
+import warnings
+from typing import TYPE_CHECKING, Final, final
 
 from django.core.cache import DEFAULT_CACHE_ALIAS, BaseCache, caches
+from django.core.exceptions import ImproperlyConfigured
 from typing_extensions import override
 
 from dmr.settings import default_parser, default_renderer
@@ -19,6 +21,48 @@ if TYPE_CHECKING:
     from dmr.throttling.algorithms import BaseThrottleAlgorithm
 
 
+@final
+class UnsafeCacheBackendWarning(UserWarning):
+    """Warning emitted when an unsafe cache backend is used for throttling."""
+
+
+_UNSAFE_CACHE_BACKENDS: Final = frozenset((
+    'django.core.cache.backends.locmem.LocMemCache',
+    'django.core.cache.backends.dummy.DummyCache',
+))
+
+_WARNING_MSG: Final = (
+    "Throttling is using '{backend}' cache backend which is not safe for "
+    'production: counters are NOT shared between processes/instances. '
+    'Use Redis or Memcached instead.'
+)
+
+_ERROR_MSG: Final = (
+    _WARNING_MSG
+    + ' To suppress this error and run at your own risk, set '
+    + "'throttle_allow_unsafe_cache': True in DMR_SETTINGS."
+)
+
+
+def _check_throttle_cache_safety(cache_name: str) -> None:
+    cache = caches[cache_name]
+    backend = f'{type(cache).__module__}.{type(cache).__qualname__}'
+
+    if backend not in _UNSAFE_CACHE_BACKENDS:
+        return
+
+    from dmr.settings import Settings, resolve_setting  # noqa: PLC0415
+
+    allow_unsafe = resolve_setting(Settings.throttle_allow_unsafe_cache)
+    if allow_unsafe:
+        warnings.warn(  # noqa: B028
+            _WARNING_MSG.format(backend=backend),
+            category=UnsafeCacheBackendWarning,
+        )
+    else:
+        raise ImproperlyConfigured(_ERROR_MSG.format(backend=backend))
+
+
 @dataclasses.dataclass(slots=True, frozen=True)
 class _DjangoCache:
     cache_name: str = DEFAULT_CACHE_ALIAS
@@ -28,11 +72,7 @@ class _DjangoCache:
         self,
     ) -> None:
         """Initialize the cache backend."""
-        from dmr.throttling.backends._cache_safety import (  # noqa: PLC0415
-            check_throttle_cache_safety,
-        )
-
-        check_throttle_cache_safety(self.cache_name)
+        _check_throttle_cache_safety(self.cache_name)
         object.__setattr__(self, '_cache', caches[self.cache_name])
 
     def _load_cache(
